% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_asymptotic_variance.R
\name{calculate_asymptotic_variance}
\alias{calculate_asymptotic_variance}
\title{Calculate Asymptotic Variance Matrix}
\usage{
calculate_asymptotic_variance(SNP, rho)
}
\arguments{
\item{SNP}{n by p matrix, genotype data for n individuals of p SNPs.}

\item{rho}{(n+1) by (n+1) matrix, sample correlation matrix of n SNPs and X.}
}
\value{
n by n matrix, which is the asymptotic covariance matrix of
sample correlations of n SNPs with X.
}
\description{
Calculate the asymptotic covariance matrix of sample correlations of n
SNPs with X, Use Theorem 2 in
"The Asymptotic Variance Matrix of the Sample Correlation Matrix".
}
