% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_3_methods_Independent.R
\name{CD_3_methods_Independent}
\alias{CD_3_methods_Independent}
\title{Perform 3 CD Methods with Multiple Independent SNPs}
\usage{
CD_3_methods_Independent(sig_part, num_iteration = 20)
}
\arguments{
\item{sig_part}{is a data.frame for the summary statistics, p by 12,
columns are: chr, pos, rsid, A1, A2, beta_T1, se_T1, N_T1, beta_T2, se_T2, N_T2, loci.}

\item{num_iteration}{number of iteration, default is 20.}
}
\value{
a list with 3 elements "CD_Ratio_result", "CD_Egger_result",
"CD_GLS_result".
}
\description{
Perform CD-Ratio, CD-Egger and CD-GLS method for two traits T1 and T2,
only calculate asymptotic covariance matrices of sample correlations once to
save time.
}
