% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Egger.R
\name{CD_Egger}
\alias{CD_Egger}
\title{Perform CD-Egger Method with Multiple SNPs}
\usage{
CD_Egger(pruned, num_iteration = 20, V_T1 = NULL, V_T2 = NULL)
}
\arguments{
\item{pruned}{a list, has "loci_bed" and "sig_part".
"loci_bed" is the reference panel, n by p matrix, genotype of
n individuals for p SNPs;
sig_part is a data.frame for the summary statistics, p by 12,
columns are "chr,pos,rsid,A1,A2,beta_T1,se_T1,N_T1,beta_T2,se_T2,N_T2,loci".}

\item{num_iteration}{number of iteration, default is 20.}

\item{V_T1}{asymptotic covariance matrix for correlations of p SNPs with T1,
default is NULL.}

\item{V_T2}{asymptotic covariance matrix for correlations of p SNPs with T2,
default is NULL.}
}
\value{
a list with four elements "T1toT2", "T2toT1", "Q_T1toT2" and
"Q_T2toT1". "T1toT2" is a vector of 4 elements: estimated K and se,
estimated intercept b0 and se, for
direction T1 to T2.
"Q_T1toT2" is the test statistic for Goodness-of-Fit Tests for direction
T1 to T2. "T2toT1" and "Q_T2toT1" are similar.
}
\description{
Perform CD-Egger method for two traits T1 and T2.
}
