% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Ratio_Independent.R
\name{CD_Ratio_Independent}
\alias{CD_Ratio_Independent}
\title{Perform CD-Ratio Method with Multiple Independent SNPs}
\usage{
CD_Ratio_Independent(sig_part, V_T1 = NULL, V_T2 = NULL)
}
\arguments{
\item{sig_part}{is a data.frame for the summary statistics, p by 12,
columns are: chr, pos, rsid, A1, A2, beta_T1, se_T1, N_T1, beta_T2, se_T2, N_T2, loci.}

\item{V_T1}{asymptotic covariance matrix for correlations of p SNPs with T1,
default is NULL.}

\item{V_T2}{asymptotic covariance matrix for correlations of p SNPs with T2,
default is NULL.}
}
\value{
a list with four elements "T1toT2", "T2toT1", "Q_T1toT2" and
"Q_T2toT1". "T1toT2" is a vector of 2 elements: estimated K and se for
direction T1 to T2.
"Q_T1toT2" is the test statistic for Goodness-of-Fit Tests for direction
T1 to T2. "T2toT1" and "Q_T2toT1" are similar.
}
\description{
Perform CD-Ratio method for two traits T1 and T2,
assuming all SNPs are valid instruments.
}
