% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD_Ratio_SingleSNP.R
\name{CD_Ratio_SingleSNP}
\alias{CD_Ratio_SingleSNP}
\title{Perform CD-Ratio Method For with A Single SNP}
\usage{
CD_Ratio_SingleSNP(pruned)
}
\arguments{
\item{pruned}{a list, has "loci_bed" and "sig_part".
"loci_bed" is the reference panel, n by 1 matrix, genotype of
n individuals for 1 SNP;
sig_part is a data.frame for the summary statistics, 1 by 12,
columns are: chr, pos, rsid, A1, A2, beta_T1, se_T1, N_T1, beta_T2, se_T2, N_T2, loci.}
}
\value{
a list with two elements "T1toT2", "T2toT1".
"T1toT2" is a vector of 2 elements: estimated K and se for
direction T1 to T2.
"T2toT1" is similar.
}
\description{
Perform CD-Ratio method for two traits T1 and T2,
with one SNP.
}
