% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sigma2.R
\name{estimate_sigma2}
\alias{estimate_sigma2}
\title{Estimate Variance}
\usage{
estimate_sigma2(v, Vg)
}
\arguments{
\item{v}{vector of length p.}

\item{Vg}{covariance matrix.}
}
\value{
Estiamte of sigma^2.
}
\description{
Use bisection to search root, which is the MLE of sigma^2.
}
