% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M_d.R
\name{generate_M_d}
\alias{generate_M_d}
\title{Generate matrix M_d}
\usage{
generate_M_d(n)
}
\arguments{
\item{n}{dimension.}
}
\value{
M_d M_d matrix.
}
\description{
Generate matrix M_d in formula 2.13 in
"The Asymptotic Variance Matrix of the Sample Correlation Matrix".
}
