% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M_s.R
\name{generate_M_s}
\alias{generate_M_s}
\title{Generate matrix M_s}
\usage{
generate_M_s(n)
}
\arguments{
\item{n}{dimension.}
}
\value{
M_s M_s matrix.
}
\description{
Generate matrix M_s in formula 2.9 in
"The Asymptotic Variance Matrix of the Sample Correlation Matrix".
}
