% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_V.R
\name{generate_V}
\alias{generate_V}
\title{Generate matrix V}
\usage{
generate_V(SNP, rho)
}
\arguments{
\item{SNP}{n by p matrix, genotype data for n individuals of p SNPs. And it is
standardized with column means 0 and variances 1.}

\item{rho}{(n+1) by (n+1) matrix, sample correlation matrix of n SNPs and X.}
}
\value{
V V matrix.
}
\description{
Generate matrix V in formula 3.6 in
"The Asymptotic Variance Matrix of the Sample Correlation Matrix".
}
